Fix "missing XModule" error
XModule could not be found because the compiled package could not be
found in edx-platform/common/lib/xmodule. This was due to the fact that
build operations realised after a VOLUME statement are ignored if they
affect the content of this volume (although it's not mounted during
build, of course). This is not a bug, it's a feature of the VOLUME
statement :)