Local filesystem storage for ORA2 uploads
By default, open response assessment files are sent to S3. Here, we
store uploaded files with the "filesystem" backend instead. To do so, we
must install a forked version of ora2 that fixes a couple issues.

Note that the "django" backend is not very usable, as file extensions
are lost on upload.

Close #97