Fix cross-platform theme assets generation
Theme assets generation relies on a cache stored in memcache. We were
observing incorrect urls generated for theme/xblock assets whenever
assets had been generated for another platform. For instance, when
viewing an XBlock in the LMS right after running "paver update_assets"
in the cms.

To solve this problem, we cache staticfiles assets in entries with
different prefix.

Close #95