Fix mysql initialization
There may be a race condition, where the mysql container is initialized
without creating the root user. To address this, we wait until the right
message from the mysql logs. In practice, we should do even better by
running the provision.sh scripts directly inside the mysql container.

This is for issues #132 and #133