Make sure that LMS can send tasks to the CMS
In production, the ALTERNATE_WORKER_QUEUES setting is overridden by ""
(empty string). This might prevent LMS from sending tasks to the CMS. We
have not seen this issue emerge yet, but better be safe than sorry.

We must be careful not to process the tasks from the CMS, just like for
the CMS worker which does not process the tasks from the LMS.