Upgrade Travis CI dist from Trusty to Xenial
Trusty is unsupported since April 2019. Also, out of the box Trusty
ships with Python 3.4, which means that we cannot replicate the CI in a
docker container. Finally, the binaries produced in Trusty are more than
twice the size of those produced in later releases.

After testing this change, unfortunately the binary size remains too
big.

See
https://discuss.overhang.io/t/size-of-tutor-16mb-built-on-my-machine-is-less-than-the-official-tutor-binary-37mb/714