fix: Only ignore warnings if they exist.
If we are running a newer version of DRF or are in the process of
upgrading DRF, we want to be able to drop the version where these
warnings exist without breaking tutor.  This pattern will suppress the
warnings if it can but will not fail to startup if that kind of warning
no longer exists.