refactor: clean mounted data folder in lms/cms
Currently, the bind-mounted lms/data and cms/data folders are a mess
because new folders are created there for every new course organisation.
These folders are empty. As far as we know they are useless... With this
change we move these folders to a dedicated "modulestore" subdirectory;
which corresponds better to the initial intent of the fs_root setting.