fix: Disable rolling upgrades for Caddy Kubernetes deployment
Caddy is a single-instance stateful application, in that it uses a
PersistentVolume that can only ever be mounted to one Pod. As such, it
is unsuitable for rolling upgrades. In the course of an upgrade, we
first need to terminate the existing Pod, then bring up its
replacement (just like the other applications configured with
PersistentVolumeClaims, like MySQL and MongoDB).

Streamline the caddy deployment with all others that use PVCs, and set
its strategy to the Recreate type.

Reference: https://kubernetes.io/docs/tasks/run-application/run-single-instance-stateful-application/