fix: incorrect redirect uri on login from cms without https
When https is disabled, the https scheme is nonetheless forced for
redirect_uri. This causes an "incorrect redirect_uri" error when login in from
the studio.
This is because of the SOCIAL_AUTH_REDIRECT_IS_HTTPS = True setting that is
defined in edx/auth-backends. This settings does not cause trouble in development
because the devstack settings explicitely set SOCIAL_AUTH_REDIRECT_IS_HTTPS =
False.